Obstacle class added
This adds an obstacle class that can spawn new obstacles (generating HTML with it) to interact with the user. Each object can have random properties assigned to it, such as width and height.

These are added to an obstacles array. Collisions and updates occur to all objects within the obstacles array now.

Some functionality on a global scope has been encapsulated within the classes themselves for cleaner logic. (e.g. bounding box, and life span of an object)